Switched to beeradvocate dataset with beer reviews.
After finding the appropriate dataset, I managed to clean up the given
file. Size of the file is 1.5GB. Because it is hard to work with such a
large amount of data I split it to chunks of 100MB for the purpose of
testing the algorithm. I changed the structure of the file using
google-refine to .csv and then managed to insert .csv data to mongodb
database.