Frontal attack
Still trying to make cemerick/friend to work with congomongo, but
without success. Have done some things to bypass difficulties in
insertion of elements, but deletion is hard  because of a nested
documents in the mongodb (requirement for friend) in user collection.
Also namespaced keywords simply cannot work with congomongo.